+++ lib/widgets/common/schedule/simple_schedule_info.dart
